pyproject.toml
[tool.poetry]
name = "backend_project"
version = "0.1.0"
description = ""
authors = ["ske-kr <54828276+ske-kr@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^3.2.9"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
